+++ src/dynamics/solver/joint_constraint/joint_velocity_constraint.rs
@@ -332,13 +332,13 @@ impl JointTwoBodyConstraint<Real, 1> {
     }
 
     pub fn solve(&mut self, solver_vels: &mut [SolverVel<Real>]) {
-        let mut solver_vel1 = solver_vels[self.solver_vel1[0] as usize];
-        let mut solver_vel2 = solver_vels[self.solver_vel2[0] as usize];
+        let mut solver_vel1 = solver_vels[self.solver_vel1[0]];
+        let mut solver_vel2 = solver_vels[self.solver_vel2[0]];
 
         self.solve_generic(&mut solver_vel1, &mut solver_vel2);
 
-        solver_vels[self.solver_vel1[0] as usize] = solver_vel1;
-        solver_vels[self.solver_vel2[0] as usize] = solver_vel2;
+        solver_vels[self.solver_vel1[0]] = solver_vel1;
+        solver_vels[self.solver_vel2[0]] = solver_vel2;
     }
 
     pub fn writeback_impulses(&self, joints_all: &mut [JointGraphEdge]) {
@@ -399,29 +399,21 @@ impl JointTwoBodyConstraint<SimdReal, SIMD_WIDTH> {
 
     pub fn solve(&mut self, solver_vels: &mut [SolverVel<Real>]) {
         let mut solver_vel1 = SolverVel {
-            linear: Vector::from(gather![
-                |ii| solver_vels[self.solver_vel1[ii] as usize].linear
-            ]),
-            angular: AngVector::from(gather![
-                |ii| solver_vels[self.solver_vel1[ii] as usize].angular
-            ]),
+            linear: Vector::from(gather![|ii| solver_vels[self.solver_vel1[ii]].linear]),
+            angular: AngVector::from(gather![|ii| solver_vels[self.solver_vel1[ii]].angular]),
         };
         let mut solver_vel2 = SolverVel {
-            linear: Vector::from(gather![
-                |ii| solver_vels[self.solver_vel2[ii] as usize].linear
-            ]),
-            angular: AngVector::from(gather![
-                |ii| solver_vels[self.solver_vel2[ii] as usize].angular
-            ]),
+            linear: Vector::from(gather![|ii| solver_vels[self.solver_vel2[ii]].linear]),
+            angular: AngVector::from(gather![|ii| solver_vels[self.solver_vel2[ii]].angular]),
         };
 
         self.solve_generic(&mut solver_vel1, &mut solver_vel2);
 
         for ii in 0..SIMD_WIDTH {
-            solver_vels[self.solver_vel1[ii] as usize].linear = solver_vel1.linear.extract(ii);
-            solver_vels[self.solver_vel1[ii] as usize].angular = solver_vel1.angular.extract(ii);
-            solver_vels[self.solver_vel2[ii] as usize].linear = solver_vel2.linear.extract(ii);
-            solver_vels[self.solver_vel2[ii] as usize].angular = solver_vel2.angular.extract(ii);
+            solver_vels[self.solver_vel1[ii]].linear = solver_vel1.linear.extract(ii);
+            solver_vels[self.solver_vel1[ii]].angular = solver_vel1.angular.extract(ii);
+            solver_vels[self.solver_vel2[ii]].linear = solver_vel2.linear.extract(ii);
+            solver_vels[self.solver_vel2[ii]].angular = solver_vel2.angular.extract(ii);
         }
     }
 
@@ -682,9 +674,9 @@ impl JointOneBodyConstraint<Real, 1> {
     }
 
     pub fn solve(&mut self, solver_vels: &mut [SolverVel<Real>]) {
-        let mut solver_vel2 = solver_vels[self.solver_vel2[0] as usize];
+        let mut solver_vel2 = solver_vels[self.solver_vel2[0]];
         self.solve_generic(&mut solver_vel2);
-        solver_vels[self.solver_vel2[0] as usize] = solver_vel2;
+        solver_vels[self.solver_vel2[0]] = solver_vel2;
     }
 
     pub fn writeback_impulses(&self, joints_all: &mut [JointGraphEdge]) {
@@ -751,19 +743,15 @@ impl JointOneBodyConstraint<SimdReal, SIMD_WIDTH> {
 
     pub fn solve(&mut self, solver_vels: &mut [SolverVel<Real>]) {
         let mut solver_vel2 = SolverVel {
-            linear: Vector::from(gather![
-                |ii| solver_vels[self.solver_vel2[ii] as usize].linear
-            ]),
-            angular: AngVector::from(gather![
-                |ii| solver_vels[self.solver_vel2[ii] as usize].angular
-            ]),
+            linear: Vector::from(gather![|ii| solver_vels[self.solver_vel2[ii]].linear]),
+            angular: AngVector::from(gather![|ii| solver_vels[self.solver_vel2[ii]].angular]),
         };
 
         self.solve_generic(&mut solver_vel2);
 
         for ii in 0..SIMD_WIDTH {
-            solver_vels[self.solver_vel2[ii] as usize].linear = solver_vel2.linear.extract(ii);
-            solver_vels[self.solver_vel2[ii] as usize].angular = solver_vel2.angular.extract(ii);
+            solver_vels[self.solver_vel2[ii]].linear = solver_vel2.linear.extract(ii);
+            solver_vels[self.solver_vel2[ii]].angular = solver_vel2.angular.extract(ii);
         }
     }
